+++ ActiveLookSDK/src/main/java/com/activelook/activelooksdk/core/ble/DiscoveredGlassesImpl.java
@@ -109,9 +109,27 @@ class DiscoveredGlassesImpl implements DiscoveredGlasses {
     ) {
         final SdkImpl sdk = BleSdkSingleton.getInstance();
         final Consumer<GlassesImpl> updater = g -> {
-            sdk.update(this, g, onConnected);
+            g.lockConnection();
+            sdk.registerConnectedGlasses(g);
+            sdk.update(this, g,
+                    g2 -> { g.unlockConnection(); onConnected.accept(g2); },
+                    onConnectionFail
+            );
         };
-        sdk.registerConnectedGlasses(new GlassesImpl(this, updater, onConnectionFail, onDisconnected));
+        new GlassesImpl(this, updater, onConnectionFail, onDisconnected);
+    }
+
+    /**
+     * Cancel previously initiated glasses connection.
+     */
+    @Override
+    public void cancelConnection() {
+        final SdkImpl sdk = BleSdkSingleton.getInstance();
+        final GlassesImpl glasses = sdk.getConnectedBleGlasses(this.getAddress());
+        if (glasses == null) {
+            throw new UnsupportedOperationException("Already disconnected.");
+        }
+        glasses.disconnect();
     }
 
     @Override